+++ chartlets.js/src/lib/components/Typography.tsx
@@ -1,9 +1,9 @@
 import MuiTypography from "@mui/material/Typography";
 import { type TypographyVariant } from "@mui/material";
 
-import { Children } from "@/lib/component/Children";
 import type { ComponentState } from "@/lib/types/state/component";
 import type { ComponentProps } from "@/lib/component/Component";
+import { Children } from "@/lib/component/Children";
 
 interface TypographyState extends ComponentState {
   align?: "right" | "left" | "center" | "inherit" | "justify";